+++ VulkanRenderer/main.cpp
@@ -14,15 +14,15 @@ int main() {
     vulkanManager.initVulkan(windowManager.getWindowRef());
     guiHandler.initImGui(windowManager.getWindowRef(), &vulkanManager);
 
-    float currTime = 0, prevTime = 0, dt = 0;
+    double currTime = 0, prevTime = 0, dt = 0;
     while (!windowManager.shouldCloseWindow()) {
         windowManager.pollWindowEvents();
 
-        vulkanManager.drawFrame(windowManager.getWindowRef(), dt, guiHandler.getReloadGUI_Flag());
+        vulkanManager.drawFrame(windowManager.getWindowRef(), static_cast<float>(dt), guiHandler.getReloadGUI_Flag());
 
         guiHandler.checkGUI_State(windowManager.getWindowRef(), &vulkanManager);
 
-        currTime = static_cast<float>(glfwGetTime());
+        currTime = glfwGetTime();
         dt = (currTime - prevTime) * 1000.0f;
         prevTime = currTime;
     }
